GIANT ASS UPDATE! ScriptingSystem, EventSystem, and Bears, oh my!
Added a ton of things, added an Event System to be used primarily by the
Scripting System, which reads in Script files and parses the text of
said scripts. Also made a couple changes in CGame, mostly to test the
Script file out, but also to Read the Inputs for the InputManager. Yay
updates and one big chunk of code in the game!